refactor class hierarchy wip
