+++ node_launcher/gui/components/cli/cli.py
