--- node_launcher/gui/network_buttons/__init__.py
@@ -1,2 +1,0 @@
-from node_launcher.gui.network_buttons.advanced.zap_layout import ZapLayout
-from .restart_layout import RestartLayout
