+++ node_launcher/gui/system_tray_widgets/__init__.py
@@ -1,0 +1,1 @@
+from .restart_layout import RestartLayout
