+++ node_launcher/gui/system_tray_widgets/bitcoind_output_widget.py
