+++ node_launcher/gui/system_tray_widgets/lnd_output_widget.py
