+++ node_launcher/gui/system_tray_widgets/restart_layout.py
