+++ node_launcher/gui/system_tray_widgets/settings/bitcoin_tab.py
@@ -2,8 +2,7 @@ from PySide2.QtCore import Qt, Signal
 from PySide2.QtWidgets import QWidget, QLabel, QCheckBox, QVBoxLayout
 
 from node_launcher.constants import Network, MAINNET, TESTNET
-from node_launcher.gui.settings.data_directories.data_directory_box import \
-    DataDirectoryBox
+from .data_directories.data_directory_box import DataDirectoryBox
 from node_launcher.node_set.bitcoin import Bitcoin
