+++ node_launcher/gui/system_tray_widgets/settings/data_directories/datadir_label.py
