+++ node_launcher/gui/system_tray_widgets/settings/data_directories/prune_warning_label.py
