+++ node_launcher/gui/system_tray_widgets/settings/lnd_tab.py
@@ -1,7 +1,7 @@
 from PySide2.QtWidgets import QWidget
 
 from node_launcher.gui.components.grid_layout import QGridLayout
-from node_launcher.gui.settings.alias_layout import AliasLayout
+from node_launcher.gui.menu_widgets.settings.alias_layout import AliasLayout
 from node_launcher.node_set.lnd import Lnd
